feat: add usage global style, bb-20
