+++ apps/mobile/src/libs/components/background-wrapper/libs/maps/size-to-styles.ts
@@ -4,6 +4,7 @@ const sizeToStyles = {
 	lg: { borderRadius: 160, height: 160, width: 160 } satisfies ViewStyle,
 	md: { borderRadius: 100, height: 100, width: 100 } satisfies ViewStyle,
 	sm: { borderRadius: 86, height: 86, width: 86 } satisfies ViewStyle,
+	xs: { borderRadius: 20, height: 33, width: 33 } satisfies ViewStyle,
 };
 
 export { sizeToStyles };
