+++ apps/mobile/src/libs/components/link/styles.ts
@@ -2,7 +2,7 @@ import { StyleSheet } from "react-native";
 
 import { GradientColor } from "~/libs/enums/enums";
 
-const COLOR_INDEX = 0;
+const COLOR_INDEX = 1;
 const styles = StyleSheet.create({
 	link: {
 		color: GradientColor.BLUE[COLOR_INDEX],
