+++ apps/mobile/src/screens/auth/components/sign-in-form/libs/constants.ts
@@ -1,0 +1,8 @@
+import { type UserSignInRequestDto } from "~/packages/users/users";
+
+const USER_SIGN_IN_DEFAULT_VALUES: UserSignInRequestDto = {
+	email: "",
+	password: "",
+};
+
+export { USER_SIGN_IN_DEFAULT_VALUES };
