+++ apps/mobile/src/screens/auth/components/sign-in-form/sign-in-form.tsx
@@ -1,20 +1,12 @@
 import React from "react";
-import { StyleSheet } from "react-native";
-//temp
 
-import {
-	BackgroundWrapper,
-	Button,
-	Input,
-	Link,
-	Text,
-	View,
-} from "~/libs/components/components";
+import { Button, Input, Link, Text, View } from "~/libs/components/components";
 import { RootScreenName } from "~/libs/enums/enums";
 import { useAppForm, useCallback } from "~/libs/hooks/hooks";
+import { globalStyles } from "~/libs/styles/styles";
 import { userSignInValidationSchema } from "~/packages/users/libs/validation-schemas/validation-schemas";
 import { type UserSignInRequestDto } from "~/packages/users/users";
-import { USER_SIGN_UP_DEFAULT_VALUES } from "~/screens/auth/components/sign-up-form/libs/constants";
+import { USER_SIGN_IN_DEFAULT_VALUES } from "~/screens/auth/components/sign-in-form/libs/constants";
 
 type Properties = {
 	onSubmit: (payload: UserSignInRequestDto) => void;
@@ -22,7 +14,7 @@ type Properties = {
 
 const SignInForm: React.FC<Properties> = ({ onSubmit }) => {
 	const { control, errors, handleSubmit } = useAppForm<UserSignInRequestDto>({
-		defaultValues: USER_SIGN_UP_DEFAULT_VALUES,
+		defaultValues: USER_SIGN_IN_DEFAULT_VALUES,
 		validationSchema: userSignInValidationSchema,
 	});
 
@@ -31,11 +23,11 @@ const SignInForm: React.FC<Properties> = ({ onSubmit }) => {
 	}, [handleSubmit, onSubmit]);
 
 	return (
-		<BackgroundWrapper>
+		<>
 			<Text preset="heading" size="xl">
 				SIGN IN
 			</Text>
-			<Text style={styles.noAccountText}>
+			<Text style={[globalStyles.mb16]}>
 				No account? Go to{" "}
 				<Link label="Create an account" to={`/${RootScreenName.SIGN_UP}`} />
 			</Text>
@@ -54,22 +46,17 @@ const SignInForm: React.FC<Properties> = ({ onSubmit }) => {
 				placeholder="********"
 			/>
 			<Button label="SIGN IN" onPress={handleFormSubmit} />
-			<View style={styles.forgotPasswordContainer}>
+			<View
+				style={[
+					globalStyles.alignItemsCenter,
+					globalStyles.justifyContentCenter,
+					globalStyles.mt16,
+				]}
+			>
 				<Link label="Forgot password?" to="/" />
 			</View>
-		</BackgroundWrapper>
+		</>
 	);
 };
-
-const styles = StyleSheet.create({
-	forgotPasswordContainer: {
-		alignItems: "center",
-		justifyContent: "center",
-		marginTop: 20,
-	},
-	noAccountText: {
-		marginBottom: 20,
-	},
-});
 
 export { SignInForm };
