+++ apps/mobile/src/screens/auth/styles.ts
@@ -1,0 +1,14 @@
+import { StyleSheet } from "react-native";
+
+import { BaseColor } from "~/libs/enums/app/base-color.enum";
+
+const styles = StyleSheet.create({
+	formContainer: {
+		backgroundColor: BaseColor.BG_WHITE,
+		borderRadius: 25,
+		height: "80%",
+		width: "90%",
+	},
+});
+
+export { styles };
